backend/flask-backend/requirements.txt
# Core Backend
Flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0
flask
# Utilities
scikit-learn==1.4.2
python-dotenv==1.0.1
requests==2.32.3

# Gemini / LLM API
google-generativeai==0.7.2
flask-pymongo==2.3.0
pymongo==4.6.1
redis==5.0.1
kafka-python==2.0.2
PyJWT==2.11.0
Pillow==10.2.0
boto3




# # Image Processing
# opencv-python==4.9.0.80
# Pillow==10.2.0
# numpy==1.26.4

# # Deep Learning (YOLOv8 compatible)
# torch==2.1.2
# torchvision==0.16.2
# torchaudio==2.1.2

# # YOLOv8
# ultralytics==8.3.30